Starting to add a new ensemble method based on the training trajectory
With this method, I generate the ensemble by taking the models
exported during the training from epoch 30,000 to 40,000, take every
100 epochs. This gives me 100 models. I decided to start at epoch
30,000 because the training loss plateaus already beyond this epoch.

The notebooks to compute the predictions are copies of the notebooks
from the other methods. The only one unique to this folder is the
script to get the ensemble models and install them.