Start separating the calculations for training NN model
Later, I want to run a bunch of trainings with different settings that
will be used by several methods. So, it makes sense to start a
training folder exclusive to store all the training calculations.

I also updated the script to use the new format of settings. I still
need to update the scripts in the other directories to do the same.