Move settings_list to settings dictionary and add another settings
The newly added settings uses batch size 1000, but learning rate is 10
times the learning rates for the other settings.

From Pratik's paper, the SGD trajectory approaches an MCMC, where the
sampling temperature is proportional to the ratio of the learning rate
over the batch size. So, by using 10 times the learning rate, the
trajectory of the training using batch size 1000 should be similar to
the trajectory using batch size 100 and the regular learning rates.