UPDATE 2025.05.01 - Scripts and notebook to investigate saturation
This investigation is to understand why the uncertainty decreases when
we extrapolate too far. Amit suggested maybe it is because the input
to the last layer are saturated, especially since we use tanh
activation function.

However, Mark doesn't seem to agree, and we can kind of see why from
the plots generated in the jupyter notebook, because even the
predictions around the equilibrium are also saturated.